animator: remodel/refactor
The previous tile-path animator implementation was incorrect.

It calculated the tile path out of order--at event processing time, rather than
in animation order.

The original intent of this organization was to avoid expensive path
recalculation for every animation frame.

To properly facilitate this, the model now includes a new "animation-init",
which creates a closure for a delta-fn.